requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
coverage==7.6.1
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
idna==3.8
iniconfig==2.0.0
inject==5.2.1
mailersend==0.5.8
packaging==24.1
pluggy==1.5.0
pytest==8.3.2
pytest-django==4.8.0
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.2
